Disable og_field_permissions.
OG Field Permissions can't take into account a non-member posting
into a group that allows for non-member postings (moderated or not).

If, as they probably are, field level permissions are required -
beyond those already in radar_og, then the same two hooks should be
implemented in radar_og, but checking for og context first and using
context if available for a node creation to test if the user can
post into fields.