@todo Just excluding radar_event_ical from not being able to export og_group_ref instance for now.
Inherited from commons is the powerful, and useful, ability to control
all group and group content fields collectively. This works nicely
with the additional request field too.

However the radar_event_ical is an old school, just post it into your
group content type and field instance.

Options
 continue to edge case this.
 work out how to exclude content types that are og, but not defined as
such by the hook for radar, from the pipe excludison.
 or integrate this form into the module too, and have an additional
value in the hook that declares it as radar og content.