Gives view _any_ eck entity privs to all users. Bug workaround.
The permission to view any entity:-

When services is checking permissions to access a entity reference field,
the id that it gets is the UUID. An entity_load is done to check permissions
to view that entity, no entity is loaded as it has the UUID not the
Drupal internal ID. The access check is then done generically.
For Taxonmy Terms for example this works as the taxonomy term access check
falls to 'access content'.
For ECK this checks for 'view all entities'.

This is deep in property_access_filter(), where it is called on the
items in a list. They have been loaded already at some point on the entity,
maybe these could be used in preference, or
where the fix should be here I've no idea yet: uuid not replacing so early,
replacing entity load, replacing the entity controller->load I dunno.

But it does actually load the node later; what's the difference?
Because it uses get_resource_reference, which does use entity_get_uuid_by_id